requirements.txt
allure-pytest==2.13.1
allure-python-commons==2.13.1
attrs==23.1.0
certifi==2022.12.7
charset-normalizer==3.1.0
exceptiongroup==1.1.1
idna==3.4
iniconfig==2.0.0
packaging==23.1
pluggy==1.0.0
py==1.11.0
pytest==7.3.1
pytest-html==3.2.0
pytest-metadata==2.0.4
requests==2.28.2
tomli==2.0.1
urllib3==1.26.15
